+++ group/views.py
@@ -11,10 +11,12 @@ from airone.lib.http import http_get, http_post, http_file_upload
 from airone.lib.http import render
 from airone.lib.http import get_download_response
 from airone.lib.http import check_superuser
+from airone.lib.profile import airone_profile
 
 from user.models import User
 
 
+@airone_profile
 @http_get
 def index(request):
     context = {}
@@ -27,6 +29,7 @@ def index(request):
     return render(request, 'list_group.html', context)
 
 
+@airone_profile
 @http_get
 @check_superuser
 def edit(request, group_id):
@@ -61,6 +64,7 @@ def edit(request, group_id):
     return render(request, 'edit_group.html', context)
 
 
+@airone_profile
 @http_post([
     {'name': 'name', 'type': str, 'checker': lambda x: x['name']},
     {'name': 'users', 'type': list, 'checker': lambda x: (
@@ -100,6 +104,7 @@ def do_edit(request, group_id, recv_data):
     })
 
 
+@airone_profile
 @http_get
 @check_superuser
 def create(request):
@@ -125,6 +130,7 @@ def create(request):
     return render(request, 'edit_group.html', context)
 
 
+@airone_profile
 @http_post([
     {'name': 'name', 'type': str, 'checker': lambda x: (
         x['name'] and not Group.objects.filter(name=x['name']).exists()
@@ -146,6 +152,7 @@ def do_create(request, recv_data):
     })
 
 
+@airone_profile
 @http_post()
 @check_superuser
 def do_delete(request, group_id, recv_data):
@@ -165,6 +172,7 @@ def do_delete(request, group_id, recv_data):
     return JsonResponse(ret)
 
 
+@airone_profile
 @http_get
 def export(request):
     user = User.objects.get(id=request.user.id)
@@ -195,11 +203,13 @@ def export(request):
     return get_download_response(output, 'user_group.yaml')
 
 
+@airone_profile
 @http_get
 def import_user_and_group(request):
     return render(request, 'import_user_and_group.html', {})
 
 
+@airone_profile
 @http_file_upload
 @check_superuser
 def do_import_user_and_group(request, context):