+++ CHANGELOG.md
@@ -7,7 +7,13 @@
 
 ### Fixed
 * Fixed not being redirected to the original URL after login
+* Fixed some request logs not output
 * Fixed the log message was not output to django.log
+
+## v2.8.0
+
+### Added
+* Added password-reset feature
 
 ## v2.7.0